+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.13.3 - 2015-01-29
+ * no longer displays Display Name fields and info if the app is not social.
+
+## 0.13.2 - 2015-01-29
+ * Button text to visit Patient Dashboard was renamed to Patient Dashboard from Patients
+
 ## 0.13.1 - 2015-01-28
  * updated membership creation view
  * refactored stylesheets